move NanoFlann & NearestNeighbor module to core::nns namespace
