+++ cpp/open3d/core/CMakeLists.txt
@@ -60,9 +60,9 @@ set(CORE_CUDA_SRC
     MemoryManagerCUDA.cu
 )
 
-set(CORE_NN_SRC
-    NanoFlann.cpp
-    NearestNeighbor.cpp
+set(CORE_NNS_SRC
+    nns/NanoFlann.cpp
+    nns/NearestNeighbor.cpp
 )
 
 if(BUILD_CUDA_MODULE)
@@ -73,14 +73,14 @@ if(BUILD_CUDA_MODULE)
         ${LINALG_CUDA_SRC}
         ${KERNEL_SRC}
         ${KERNEL_CUDA_SRC}
-        ${CORE_NN_SRC}
+        ${CORE_NNS_SRC}
     )
 else()
   set(ALL_CORE_SRC
         ${CORE_SRC}
         ${LINALG_SRC}
         ${KERNEL_SRC}
-        ${CORE_NN_SRC}
+        ${CORE_NNS_SRC}
     )
 endif()
